// frontend/src/features/Mission/mission.type.ts
export namespace Mission {
  export const CIRCULAR_ZONE_RADIUS = 1500
  export enum ActionTypeEnum {
    CONTROL = 'CONTROL',
    DETACHED_REPORTING = 'DETACHED_REPORTING',
    NOTE = 'NOTE',
    REPORTING = 'REPORTING',
    SURVEILLANCE = 'SURVEILLANCE'
  }
  export const actionTypeLabels = {
    CONTROL: {
      code: 'CONTROL',
      libelle: 'Contrôle'
    },
    NOTE: {
      code: 'NOTE',
      libelle: 'Note'
    },
    SURVEILLANCE: {
      code: 'SURVEILLANCE',
      libelle: 'Surveillance'
    }
  }

  export enum MissionTypeEnum {
    AIR = 'AIR',
    LAND = 'LAND',
    SEA = 'SEA'
  }
  export const missionTypeLabels = {
    SEA: {
      code: 'SEA',
      libelle: 'Mer'
    },
    // eslint-disable-next-line sort-keys-fix/sort-keys-fix
    LAND: {
      code: 'LAND',
      libelle: 'Terre'
    },
    // eslint-disable-next-line sort-keys-fix/sort-keys-fix
    AIR: {
      code: 'AIR',
      libelle: 'Air'
    }
  }

  export enum CompletionStatus {
    COMPLETED = 'COMPLETED',
    TO_COMPLETE = 'TO_COMPLETE'
  }
// ...
  export enum InfractionTypeEnum {
    WAITING = 'WAITING',
    WITHOUT_REPORT = 'WITHOUT_REPORT',
    WITH_REPORT = 'WITH_REPORT'
  }
  export const infractionTypeLabels = {
    WITH_REPORT: {
      code: 'WITH_REPORT',
      libelle: 'Avec PV'
    },
    WITHOUT_REPORT: {
      code: 'WITHOUT_REPORT',
      libelle: 'Sans PV'
    },
    // eslint-disable-next-line sort-keys-fix/sort-keys-fix
    WAITING: {
      code: 'WAITING',
      libelle: 'En attente'
    }
  }

  export enum FormalNoticeEnum {
    NO = 'NO',
    PENDING = 'PENDING',
    YES = 'YES'
  }
// ...
  export enum VesselTypeEnum {
    COMMERCIAL = 'COMMERCIAL',
    FISHING = 'FISHING',
    MOTOR = 'MOTOR',
    SAILING = 'SAILING'
  }
// ...
  export enum MissionSourceEnum {
    MONITORENV = 'MONITORENV',
    MONITORFISH = 'MONITORFISH',
    POSEIDON_CACEM = 'POSEIDON_CACEM',
    POSEIDON_CNSP = 'POSEIDON_CNSP'
  }
// ...
  export enum ActionSource {
    MONITORENV = 'MONITORENV',
    MONITORFISH = 'MONITORFISH'
  }

  // Mission from API
  export type Mission<EnvAction = EnvActionControl | EnvActionSurveillance | EnvActionNote> = {
    attachedReportingIds: number[]
    attachedReportings: ReportingDetailed[]
    completedBy: string
    controlUnits: LegacyControlUnit[]
    createdAtUtc?: string | undefined
    detachedReportingIds: number[]
    detachedReportings?: []
    endDateTimeUtc?: string
    envActions: EnvAction[]
    facade: SeaFrontEnum
    fishActions: FishMissionAction.MissionAction[]
    geom?: GeoJSON.MultiPolygon
    hasMissionOrder?: boolean
    id: number
    isGeometryComputedFromControls: boolean
    isUnderJdp?: boolean
    missionSource: MissionSourceEnum
    missionTypes: MissionTypeEnum[]
    observationsCacem?: string
    observationsCnsp?: string
    openBy: string
    startDateTimeUtc: string
    updatedAtUtc?: string | undefined
  }

  export type NewMission = Omit<Mission<NewEnvAction>, 'controlUnits' | 'facade' | 'id'> & {
    controlUnits: Array<Omit<LegacyControlUnit, 'administrationId' | 'id'>>
    id: string
  }
  // Mission for API
  export type MissionData = Omit<Partial<Mission<EnvAction>>, 'attachedReportings'>

  export type EnvAction = EnvActionControl | EnvActionSurveillance | EnvActionNote
  export type NewEnvAction = NewEnvActionControl | EnvActionSurveillance | EnvActionNote

  export type EnvActionCommonProperties = {
    actionStartDateTimeUtc?: string | null
    id: string
  }

  export type NewEnvActionControl = EnvActionCommonProperties & {
    actionEndDateTimeUtc?: string | null
    actionNumberOfControls?: number
    actionTargetType?: string
    actionType: ActionTypeEnum.CONTROL
    completedBy: string
    completion: CompletionStatus
    controlPlans: ControlPlansData[]
    geom?: GeoJSON.MultiPolygon | GeoJSON.MultiPoint
    infractions: Infraction[] | NewInfraction[]
    isAdministrativeControl?: boolean
    isComplianceWithWaterRegulationsControl?: boolean
    isSafetyEquipmentAndStandardsComplianceControl?: boolean
    isSeafarersControl?: boolean
    observations: string | null
    openBy: string
    reportingIds: number[]
    vehicleType?: string
  }
  export type EnvActionControl = NewEnvActionControl & {
    actionTargetType: string
  }

  export type EnvActionSurveillance = EnvActionCommonProperties & {
    actionEndDateTimeUtc?: string | null
    actionType: ActionTypeEnum.SURVEILLANCE
    completedBy: string
    completion: CompletionStatus
    controlPlans: ControlPlansData[]
    durationMatchesMission?: boolean
    geom?: GeoJSON.MultiPolygon
    observations: string | null
    openBy: string
    reportingIds: number[]
  }

  export type ControlOrSurveillance =
    | EnvActionControl
    | EnvActionSurveillance
    | NewEnvActionControl
    | EnvActionSurveillance

  export type EnvActionNote = EnvActionCommonProperties & {
    actionType: ActionTypeEnum.NOTE
    observations?: string | null
  }

  export type NewInfraction = {
    companyName?: string | null
    controlledPersonIdentity?: string | null
    formalNotice?: FormalNoticeEnum
    id: string
    infractionType?: InfractionTypeEnum
    natinf?: string[]
    observations?: string | null
    registrationNumber?: string | null
    relevantCourt?: string | null
    toProcess: boolean
    vesselSize?: Number | null
    vesselType?: VesselTypeEnum | null
  }
  export type Infraction = NewInfraction & {
    formalNotice: FormalNoticeEnum
    infractionType: InfractionTypeEnum
  }

  export type EnvActionForTimeline = EnvAction & {
    actionSource: ActionSource
    formattedReportingId: string
    formattedReportingIds: string[]
    timelineDate?: string
  }

  export type ActionsTypeForTimeLine =
    | ReportingForTimeline
    | DetachedReportingForTimeline
    | Mission.EnvActionForTimeline
    | FishMissionAction.FishActionForTimeline
}